+++ spec/unit/hanami/cli/commands/app/generate/action_spec.rb
@@ -1086,6 +1086,49 @@ RSpec.describe Hanami::CLI::Commands::App::Generate::Action, :app do
     include_context "with existing files" do
       let(:generate_action) { subject.call(name: action_name) }
     end
+
+    # This namespacing behavior is shared across all generators and lives in
+    # `Hanami::CLI::RubyFileGenerator`. This is just an integration check; see
+    # spec/unit/hanami/cli/ruby_file_generator_spec.rb for full coverage.
+    context "when the action namespace includes the app name" do
+      it "namespaces the action and the view" do
+        within_application_directory do
+          subject.call(name: "test/foo.index")
+
+          expect(fs.read("app/actions/test/foo/index.rb")).to eq(<<~EXPECTED)
+            # frozen_string_literal: true
+
+            module Test
+              module Actions
+                module Test
+                  module Foo
+                    class Index < ::Test::Action
+                      def handle(request, response)
+                      end
+                    end
+                  end
+                end
+              end
+            end
+          EXPECTED
+
+          expect(fs.read("app/views/test/foo/index.rb")).to eq(<<~EXPECTED)
+            # frozen_string_literal: true
+
+            module Test
+              module Views
+                module Test
+                  module Foo
+                    class Index < ::Test::View
+                    end
+                  end
+                end
+              end
+            end
+          EXPECTED
+        end
+      end
+    end
   end
 
   context "generate for a slice" do
@@ -1526,6 +1569,47 @@ RSpec.describe Hanami::CLI::Commands::App::Generate::Action, :app do
 
           expect(fs.exist?("spec/slices/#{slice}/actions/no/test_spec.rb")).to be(false)
           expect(output).to_not include("Created spec/slices/#{slice}/actions/no/test_spec.rb")
+        end
+      end
+
+      # This namespacing behavior is shared across all generators and lives in
+      # `Hanami::CLI::RubyFileGenerator`. This is just an integration check; see
+      # spec/unit/hanami/cli/ruby_file_generator_spec.rb for full coverage.
+      context "when the action namespace includes the slice name" do
+        it "namespaces the action and the view" do
+          within_application_directory do
+            prepare_slice!
+
+            subject.call(name: "main.index", slice: "main")
+
+            expect(fs.read("slices/main/actions/main/index.rb")).to eq(<<~EXPECTED)
+              # frozen_string_literal: true
+
+              module Main
+                module Actions
+                  module Main
+                    class Index < ::Main::Action
+                      def handle(request, response)
+                      end
+                    end
+                  end
+                end
+              end
+            EXPECTED
+
+            expect(fs.read("slices/main/views/main/index.rb")).to eq(<<~EXPECTED)
+              # frozen_string_literal: true
+
+              module Main
+                module Views
+                  module Main
+                    class Index < ::Main::View
+                    end
+                  end
+                end
+              end
+            EXPECTED
+          end
         end
       end
     end